+++ RobotTopDownV3/Assets/Scripts/Hub/StructureUpgrade.cs
@@ -5,9 +5,9 @@ public abstract class StructureUpgrade : IncrementalUpgrade
 {
 	public string[] addonDescriptions;
 
-	public abstract int GetAddonValue ( int _level, int _addonID );
+	public abstract float GetAddonValue ( int _level, int _addonID );
 
-	public string GetAddonDescription(int _addonID, int _bonus )
+	public string GetAddonDescription(int _addonID, float _bonus )
 	{
 		return "+" + _bonus + " " + addonDescriptions[_addonID];
 	}
